move fight codec and make it reset hp at fight starts
